requirements.txt
Django==1.5.5
Pillow==2.2.1
argparse==1.2.1
django-admin-sortable==1.5.5
django-ckeditor==4.0.2
django-extensions==1.2.5
ipython==1.1.0
psycopg2==2.5.1
pygraphviz==1.2
six==1.4.1
wsgiref==0.1.2
